working on coaching, must change Coaching to a DTO so user can mapped properly onto it

+++ src/main/java/codersbay/vienna/nachhilfe/wien/backend/dto/ProfileDTO.java
@@ -11,7 +11,7 @@ public class ProfileDTO {
     String email;
     String description;
     Boolean active;
-    Integer averageRatingScore;
+    Double averageRatingScore;
 
 }
